+++ src/@lib/hocs/deep-memo.ts
@@ -1,5 +1,5 @@
-import { deepEquals } from "../equalities";
 import { ComponentType } from "react";
+import { deepEquals } from "../equalities/index.ts";
 import { memo } from "./memo.ts";
 
 export function deepMemo<P extends object>(Component: ComponentType<P>) {
