+++ src/@lib/hooks/index.ts
@@ -1,4 +1,4 @@
-export * from "./useDeepMemo";
-export * from "./useMemo";
-export * from "./useCallback";
-export * from "./useRef";
+export * from "./use-callback";
+export * from "./use-deep-memo";
+export * from "./use-memo";
+export * from "./use-ref";
