+++ src/@lib/hooks/use-memo.ts
@@ -1,6 +1,6 @@
 import { DependencyList } from "react";
 import { shallowEquals } from "../equalities";
-import { useRef } from "./useRef";
+import { useRef } from "./use-ref";
 
 export function useMemo<T>(
   factory: () => T,
