+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export * from "./ComplexForm";
-export * from "./Header";
-export * from "./ItemList";
-export * from "./NotificationSystem";
+export * from "./complex-form";
+export * from "./hheader";
+export * from "./item-list";
+export * from "./notification-system";
